Move to shared test
