+++ maven-resolver-test-http/src/main/resources/ssl/README.txt
