+++ maven-resolver-test-http/src/main/resources/ssl/client-store
