+++ maven-resolver-test-http/src/main/resources/ssl/server-store-selfsigned
