--- maven-resolver-transport-wagon/src/test/resources/logback-test.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
- ! Licensed to the Apache Software Foundation (ASF) under one
- ! or more contributor license agreements.  See the NOTICE file
- ! distributed with this work for additional information
- ! regarding copyright ownership.  The ASF licenses this file
- ! to you under the Apache License, Version 2.0 (the
- ! "License"); you may not use this file except in compliance
- ! with the License.  You may obtain a copy of the License at
- ! 
- !  http://www.apache.org/licenses/LICENSE-2.0
- ! 
- ! Unless required by applicable law or agreed to in writing,
- ! software distributed under the License is distributed on an
- ! "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- ! KIND, either express or implied.  See the License for the
- ! specific language governing permissions and limitations
- ! under the License.
- !-->
-
-<configuration>
-  <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
-    <encoder>
-      <pattern>%d{HH:mm:ss.SSS} [%-18thread] %c{1} [%p] %m%n</pattern>
-    </encoder>
-  </appender>
-
-  <logger name="org.sonatype.tests.jetty.server" level="INFO"/>
-  <logger name="org.sonatype.tests.jetty.server.behaviour" level="DEBUG"/>
-  <logger name="org.sonatype.tests" level="DEBUG"/>
-
-  <root level="WARN">
-    <appender-ref ref="CONSOLE"/>
-  </root>
-</configuration>
